requirements.txt
Flask
boto3
flask-cors
pyjwt
python-dotenv
zappa
cryptography